HW02 done

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryUserMealRepositoryImpl.java
@@ -1,18 +1,24 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.springframework.stereotype.Repository;
 import ru.javawebinar.topjava.model.UserMeal;
 import ru.javawebinar.topjava.repository.UserMealRepository;
 import ru.javawebinar.topjava.util.UserMealsUtil;
+import ru.javawebinar.topjava.util.exception.NotFoundException;
 
-import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
+
+import static ru.javawebinar.topjava.LoggedUser.id;
 
 /**
  * GKislin
  * 15.09.2015.
  */
+@Repository
 public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
     private Map<Integer, UserMeal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
@@ -23,6 +29,7 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
 
     @Override
     public UserMeal save(UserMeal userMeal) {
+
         if (userMeal.isNew()) {
             userMeal.setId(counter.incrementAndGet());
         }
@@ -32,17 +39,30 @@ public class InMemoryUserMealRepositoryImpl implements UserMealRepository {
 
     @Override
     public void delete(int id) {
-        repository.remove(id);
+        if (id() == repository.get(id).getUserId()) {
+            repository.remove(id);
+        }
     }
 
     @Override
     public UserMeal get(int id) {
-        return repository.get(id);
+        UserMeal userMeal = repository.get(id);
+        if (id() == userMeal.getUserId()) {
+            return userMeal;
+        }
+
+        throw new NotFoundException();
     }
 
     @Override
-    public Collection<UserMeal> getAll() {
-        return repository.values();
+    public List<UserMeal> getAll() {
+        return
+                repository
+                        .values()
+                        .stream()
+                        .filter(userMeal -> id() == userMeal.getUserId())
+                        .sorted((o1, o2) -> o1.getDateTime().compareTo(o2.getDateTime()))
+                        .collect(Collectors.toList());
     }
 }
